+++ python/robot.cpp
@@ -985,7 +985,7 @@ Examples
 >>> robot.servo_off(".*")  # All motor drivers
 >>> # Specific device: robot.servo_off("^torso_.*")
 )doc")
-      .def("break_engage", &Robot<T>::BreakEngage, "dev_name"_a, R"doc(
+      .def("brake_engage", &Robot<T>::BrakeEngage, "dev_name"_a, R"doc(
 
 Engage the brake for a device.
 
@@ -999,7 +999,7 @@ Examples
 >>> # Engage brake for manitance
 >>> robot.break_engage("right_arm_0")
 )doc")
-      .def("break_release", &Robot<T>::BreakRelease, "dev_name"_a, R"doc(
+      .def("brake_release", &Robot<T>::BrakeRelease, "dev_name"_a, R"doc(
 break_release(dev_name)
 
 Release the brake for a device.